incorporate stacked bar chart into ef plot CLI
